+++ bundles/framework/bundle/mapmodule-plugin/request/MapMoveRequestHandler.js
@@ -20,7 +20,9 @@ function(sandbox, mapModule) {
 }, {
     /**
      * @method handleRequest 
-     * Handles the request
+     * Handles the request.
+     * If the request SrsName is not defined in Proj4js.defs then a "SrsName not supported!" exception is thrown.
+     *
      * @param {Oskari.mapframework.core.Core} core
      *      reference to the application core (reference sandbox core.getSandbox())
      * @param {Oskari.mapframework.request.common.MapMoveRequest} request
@@ -31,8 +33,19 @@ function(sandbox, mapModule) {
         var latitude = request.getCenterY();
         var marker = request.getMarker();
         var zoom = request.getZoom();
+        var srsName = request.getSrsName();
 
         var lonlat = new OpenLayers.LonLat(longitude, latitude);
+
+        // transform coordinates to given projection
+        if (srsName && (this.mapModule.getProjection() !== srsName)) {
+            var isProjectionDefined = Proj4js.defs[srsName];
+            if (!isProjectionDefined) {
+                throw "SrsName not supported!";
+            }
+            lonlat = this.mapModule.transformCoordinates(lonlat, srsName);
+        }
+
         this.mapModule.moveMapToLanLot(lonlat, zoom, false);
         // if zoom=0 -> if(zoom) is determined as false...
         if (zoom || zoom === 0) {