+++ bundles/framework/bundle/personaldata/MyPlacesTab.js
@@ -391,7 +391,7 @@ function(instance, localization) {
         var sandbox = instance.getSandbox();
         // faking to be module with getName/onEvent methods
         for (p in this.eventHandlers) {
-            sandbox.unregisterForEventByName(this, p);
+            sandbox.unregisterFromEventByName(this, p);
         }
     }
 });